Split core orchestration service boundaries
